Add validation functions
