+++ src/traits/election/static_committee.rs
@@ -6,7 +6,7 @@ use hotshot_types::{
     traits::{
         election::{Checked, Election, ElectionConfig, ElectionError, VoteToken},
         node_implementation::NodeType,
-        signature_key::{EncodedPublicKey, EncodedSignature, SignatureKey},
+        signature_key::{EncodedSignature, SignatureKey},
     },
 };
 use jf_primitives::signatures::BLSSignatureScheme;
@@ -99,39 +99,6 @@ where
     type DACertificate = LEAF::DACertificate;
 
     type LeafType = LEAF;
-
-    fn is_valid_qc(&self, _qc: &Self::QuorumCertificate) -> bool {
-        #[allow(deprecated)]
-        nll_todo()
-    }
-
-    fn is_valid_dac(&self, _qc: Self::DACertificate) -> bool {
-        #[allow(deprecated)]
-        nll_todo()
-    }
-
-    fn is_valid_qc_signature(
-        &self,
-        _encoded_key: &EncodedPublicKey,
-        _encoded_signature: &EncodedSignature,
-        _hash: Commitment<Self::LeafType>,
-        _view_number: TYPES::Time,
-        _vote_token: Checked<TYPES::VoteTokenType>,
-    ) -> bool {
-        #[allow(deprecated)]
-        nll_todo()
-    }
-
-    fn is_valid_dac_signature(
-        &self,
-        _encoded_key: &EncodedPublicKey,
-        _encoded_signature: &EncodedSignature,
-        _view_number: TYPES::Time,
-        _vote_token: Checked<TYPES::VoteTokenType>,
-    ) -> bool {
-        #[allow(deprecated)]
-        nll_todo()
-    }
 
     /// Clone the static table
     fn get_stake_table(
